Ending

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,10 +19,10 @@ class CreateUsersTable extends Migration
             $table->bigIncrements('id');
             $table->string('name');
             $table->string('email')->unique();
-            $table->integer('cedula')->nullable();
+            $table->bigInteger('cedula')->nullable();
             $table->bigInteger('telefono')->nullable();
             $table->string('direccion')->nullable();
-            $table->integer('NIT')->nullable();
+            $table->bigInteger('NIT')->nullable();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->rememberToken();
